fix(sheets-formula): preserve reference highlight editing
